chore(thermal): renaming
