+++ tests/integration/study_data_blueprint/test_thermal.py
@@ -51,7 +51,7 @@ import pytest
 from starlette.testclient import TestClient
 
 from antarest.core.utils.string import to_camel_case
-from antarest.study.business.model.thermal_model import ThermalCluster
+from antarest.study.business.model.thermal_cluster_model import ThermalCluster
 from antarest.study.model import STUDY_VERSION_8_6, STUDY_VERSION_8_7
 from antarest.study.storage.rawstudy.model.filesystem.config.identifier import transform_name_to_id
 from tests.integration.utils import wait_task_completion
