+++ tests/storage/repository/filesystem/test_scenariobuilder.py
@@ -13,7 +13,7 @@
 from pathlib import Path
 from unittest.mock import Mock
 
-from antarest.study.business.model.thermal_model import ThermalCluster
+from antarest.study.business.model.thermal_cluster_model import ThermalCluster
 from antarest.study.storage.rawstudy.model.filesystem.config.model import Area, FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.root.settings.scenariobuilder import ScenarioBuilder
